Refactored resources validation routine
