+++ datapackagist/src/scripts/components/ui/bigbang.js
@@ -22,9 +22,8 @@ module.exports = backbone.BaseView.extend({
     this.layout.download = new DownloadView({el: window.APP.$('#download-data-package')});
     this.layout.navbar = new navigation.NavbarView({el: window.APP.$('#navbar')});
 
-    this.layout.validationResultList = (new validationResult.ListView({
-      el: window.APP.$('#validation-result'),
-      container: '[data-id=list-container]'
+    this.layout.validationResultList = (new validationResult.ValidationResultsView({
+      el: window.APP.$('#validation-result')
     })).render();
 
     return this;
